refactor: MenuPrice 래핑
