+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,15 +1,14 @@
 package kitchenpos.menu.domain;
 
 import javax.persistence.CascadeType;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 @Entity
 public class Menu {
@@ -18,7 +17,8 @@ public class Menu {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private BigDecimal price;
+    @Embedded
+    private MenuPrice price;
     private Long menuGroupId;
     @OneToMany(mappedBy = "menu", cascade = {CascadeType.PERSIST})
     private List<MenuProduct> menuProducts = new ArrayList<>();
@@ -26,10 +26,7 @@ public class Menu {
     protected Menu() {
     }
 
-    public Menu(String name, BigDecimal price, Long menuGroupId) {
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException("가격이 0보다 작거나 null일 수 없습니다.");
-        }
+    public Menu(final String name, final MenuPrice price, final Long menuGroupId) {
         this.name = name;
         this.price = price;
         this.menuGroupId = menuGroupId;
