+++ src/main/java/kitchenpos/menu/domain/MenuPrice.java
@@ -1,0 +1,21 @@
+package kitchenpos.menu.domain;
+
+import javax.persistence.Embeddable;
+import java.math.BigDecimal;
+import java.util.Objects;
+
+@Embeddable
+public class MenuPrice {
+
+    private BigDecimal price;
+
+    protected MenuPrice() {
+    }
+
+    public MenuPrice(final BigDecimal price) {
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("가격이 0보다 작거나 null일 수 없습니다.");
+        }
+        this.price = price;
+    }
+}
